Update licensing to Apache-2.0 WITH LLVM-exception (#131)
This is important for use with codebases licensed as GPL-2.0-only, such as
the Linux kernel and some open-source game engines.

It also allows the user to make use of NVTX without any additional burden
of including that fact in their documentation.

Note: This change requires approval and sign-off before it takes effect.